Update FinalProject.cpp
added more options to select from. The introduction needs to be finished with a list of how to access every function. The while function also needs to be finished there needs to be ether an || our statement so Q and q or change the choice variable to uppercase. 
The choice of states also needs to be fixed the spacing issue is still there and I'm having a lot of trouble solving it.